Improve UX when user not found in 'forgot-password'
Backend responds with a 404 when a user is not found. The code in the
controller is meant to deal with it but QueryService throws an exception
on 404 instead of returning the response. The controller now catches
such 404 exceptions, and also properly presents an error message when
there is some other bad http status code.